# LaravelBlog-1.0.0-swagger.yaml
openapi: 3.0.0
...
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/ship87/LaravelBlog/1.0.0
info:
  description: This is a Laravel Blog API
  version: "1.0.0"
  title: LaravelBlog API
  contact:
    email: theship87@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
  /posts:
    get:
      tags:
        - posts
      summary: search posts
      operationId: searchPosts
      description: Searсh posts
      parameters:
        - in: query
          name: search
          description: Search text in posts
          required: false
          schema:
            type: string
        - in: query
          name: page
          description: Number of posts for pagination
          schema:
            type: integer
            format: int32
            minimum: 0
        - in: query
          name: include
          description: Add relationships
          required: false
          schema:
            type: string
        - in: query
          name: sort
          description: Sort by column
          required: false
          schema:
            type: string
        - in: query
          name: filter[column]
          description: Filter by column
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 'search posts matching criteria'
          content:
            application/vnd.api+json:
              schema:
                type: array
                items:
                    $ref: '#/components/schemas/Posts'
        '400':
          description: 'invalid input, object invalid'
    post:
      tags:
        - posts
      summary: add post
      operationId: addPost
      description: Adds an post to the blog
      responses:
        '201':
          description: 'post created'
        '400':
          description: 'invalid input, object invalid'
      requestBody:
        content:
          application/vnd.api+json:
            schema:
              $ref: '#/components/schemas/Post'
...
  /pages:
    get:
      tags:
        - pages
      summary: search pages
      operationId: searchPages
      description: Search pages
      parameters:
        - in: query
          name: search
          description: Search text in pages
          required: false
          schema:
            type: string
        - in: query
          name: page
          description: Number of pages for pagination
          schema:
            type: integer
            format: int32
            minimum: 0
        - in: query
          name: include
          description: Add relationships
          required: false
          schema:
            type: string
        - in: query
          name: sort
          description: Sort by column
          required: false
          schema:
            type: string
        - in: query
          name: filter[column]
          description: Filter by column
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 'search pages matching criteria'
          content:
            application/vnd.api+json:
              schema:
                type: array
                items:
                    $ref: '#/components/schemas/Pages'
        '400':
          description: 'invalid input, object invalid'
    post:
      tags:
        - pages
      summary: add page
      operationId: addPage
      description: Adds an page to the blog
      responses:
        '201':
          description: 'page created'
        '400':
          description: 'invalid input, object invalid'
      requestBody:
        content:
          application/vnd.api+json:
            schema:
              $ref: '#/components/schemas/Page'
...
components:
  schemas:
    Post:
      type: object
      required:
        - id
        - type
        - attributes
      properties:
          id:
            type: integer
            format: int32
            minimum: 1
            example: 1
          type:
            type: string
            example: pages
          attributes:
            type: object
            properties:
              title:
                type: string
                example: Post one
              slug:
                type: string
                example: post-one
              content:
                type: string
                example: Text of post one
              no_comment:
                type: string
                example: N
              created_user_id:
                type: integer
                format: int32
                minimum: 1
              updated_user_id:
                type: integer
                format: int32
                minimum: 1
              created_at:
                type: string
                format: date-time
                example: '2019-01-12T09:12:33.001Z'
              updated_at:
                type: string
                format: date-time
                example: '2019-02-05T07:52:22.001Z'
              deleted_at:
                type: string
                format: date-time
                example: '2019-02-05T07:52:22.001Z'
    Posts:
      type: object
      required:
        - data
        - id
        - type
        - attributes
      properties:
       data:
        $ref: '#/components/schemas/Post'
    Page:
      type: object
      required:
        - id
        - type
        - attributes
      properties:
          id:
            type: integer
            format: int32
            minimum: 1
            example: 1
          type:
            type: string
            example: pages
          attributes:
            type: object
            properties:
              title:
                type: string
                example: Post one
              slug:
                type: string
                example: post-one
              content:
                type: string
                example: Text of post one
              parent_id:
                type: integer
                format: int32
                minimum: 1
                example: 2
              no_comment:
                type: string
                example: N
              created_user_id:
                type: integer
                format: int32
                minimum: 1
              updated_user_id:
                type: integer
                format: int32
                minimum: 1
              created_at:
                type: string
                format: date-time
                example: '2019-01-12T09:12:33.001Z'
              updated_at:
                type: string
                format: date-time
                example: '2019-02-05T07:52:22.001Z'
              deleted_at:
                type: string
                format: date-time
                example: '2019-02-05T07:52:22.001Z'
    Pages:
      type: object
      required:
        - data
        - id
        - type
        - attributes
      properties:
       data:
        $ref: '#/components/schemas/Page'
...